requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyascii==0.3.2
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.3
django-encrypted-model-fields==0.6.5
django-filter==25.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.0
django-taggit==6.1.0
django-tasks==0.7.0
django-treebeard==4.7.1
django-unfold==0.60.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
frozenlist==1.7.0
g4f==0.5.5.5
idna==3.10
laces==0.1.2
multidict==6.5.0
nest-asyncio==1.6.0
openpyxl==3.1.5
pillow==11.2.1
pillow_heif==0.21.0
propcache==0.3.2
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.9.0
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wagtail==7.0.1
Willow==1.10.0
yarl==1.20.1
